*Major* misc changes
There's just too many of 'em :(

Noteable changes:

- Added invert function to bitmap class
- Add const like everywhere. Idk.
- Fixed coins' block form hit detection
- Added flying prize blocks
- Improve grab blocks (not quite perfect yet)

+++ GameSession.h
@@ -47,7 +47,7 @@ private:
 
 	static void ReadAllSessionData(); // Puts all available data into m_AllSessionInfoArr (fast, but takes some time if sessions file is large)
 	static int GetNumberOfSessions();
-	static SessionInfoPair FindSessionInfoPair(int sessionIndex); // Runs through the file string until it finds the nth session start tag (SLOW)
+	static SessionInfoPair FindSessionInfoPair(size_t sessionIndex); // Runs through the file string until it finds the nth session start tag (SLOW)
 	static SessionInfo GetSessionInfo(const std::string& sessionString);
 	static SessionInfoPair GetSessionInfoPair(const std::string& sessionString, int sessionIndex);
 
@@ -66,13 +66,13 @@ private:
 	static bool m_DaysOfWeekArrGenerated;
 	static bool m_ShouldGenerateDaysOfWeek;
 
-	static int m_CurrentSessionShowingIndex;
+	static size_t m_CurrentSessionShowingIndex;
 
 	// This should only be modified at the start and end of a game!
 	static SessionInfoPair m_CurrentSessionInfoRecording;
 	
 	static std::string m_AllSessionsInfoString;
-	static int m_NumberOfSessions;
+	static size_t m_NumberOfSessions;
 	// This gets filled one element at a time, as the user requests them, until they press F2 to request all (for a nice graph)
 	// The first element is the most recent, the last element is the oldest
 	static std::vector<SessionInfoPair> m_AllSessionInfoArr;
